Move condition to another file
