+++ models/keyword.go
@@ -3,8 +3,6 @@ package models
 import (
 	"database/sql/driver"
 	"errors"
-	"fmt"
-	"strings"
 
 	"github.com/gutakk/go-google-scraper/db"
 
@@ -25,17 +23,6 @@ const (
 	InvalidKeywordStatusErr    = "invalid keyword status"
 	couldNotJoinConditionError = "could not join conditions"
 )
-
-type Condition struct {
-	ConditionName string
-	Value         string
-}
-
-// Map query string filter to where statement
-var ConditionMapper = map[string]string{
-	"keyword": "LOWER(keyword) LIKE LOWER('%%%s%%')",
-	"user_id": "user_id = '%s'",
-}
 
 func (k KeywordStatus) Value() (driver.Value, error) {
 	switch k {
@@ -75,7 +62,7 @@ func GetKeywordBy(condition map[string]interface{}) (Keyword, error) {
 func GetKeywordsBy(conditions []Condition) ([]Keyword, error) {
 	var keywords []Keyword
 
-	joinedConditions, err := getJoinedConditions(conditions)
+	joinedConditions, err := GetJoinedConditions(conditions)
 	if err != nil {
 		return nil, err
 	}
@@ -111,24 +98,4 @@ func UpdateKeyword(keywordID uint, newKeyword Keyword) error {
 	}
 
 	return nil
-}
-
-func getJoinedConditions(conditions []Condition) (string, error) {
-	var formattedConditions []string
-
-	for _, c := range conditions {
-		conditionName := c.ConditionName
-		conditionValue := c.Value
-		whereStatement := ConditionMapper[conditionName]
-
-		if conditionValue != "" && whereStatement != "" {
-			formattedConditions = append(formattedConditions, fmt.Sprintf(whereStatement, conditionValue))
-		} else {
-			return "", errors.New(couldNotJoinConditionError)
-		}
-	}
-
-	joinedConditions := strings.Join(formattedConditions, " AND ")
-
-	return joinedConditions, nil
 }
