+++ models/keyword_test.go
@@ -312,59 +312,3 @@ func (s *KeywordDBTestSuite) TestUpdateKeywordWithInvalidStatus() {
 
 	assert.Equal(s.T(), "invalid keyword status", err.Error())
 }
-
-func TestGetJoinedConditionsWithValidConditionsMap(t *testing.T) {
-	conditions := []Condition{
-		{
-			ConditionName: "keyword",
-			Value:         "testKeyword",
-		},
-		{
-			ConditionName: "user_id",
-			Value:         "testUserID",
-		},
-	}
-
-	result, err := getJoinedConditions(conditions)
-
-	expected := "LOWER(keyword) LIKE LOWER('%testKeyword%') AND user_id = 'testUserID'"
-
-	assert.Equal(t, expected, result)
-	assert.Equal(t, nil, err)
-}
-
-func TestGetJoinedConditionsWithInvalidFilter(t *testing.T) {
-	conditions := []Condition{
-		{
-			ConditionName: "invalid",
-			Value:         "testKeyword",
-		},
-		{
-			ConditionName: "invalid",
-			Value:         "testUserID",
-		},
-	}
-
-	result, err := getJoinedConditions(conditions)
-
-	assert.Equal(t, "", result)
-	assert.Equal(t, "could not join conditions", err.Error())
-}
-
-func TestGetJoinedConditionsWithBlankValue(t *testing.T) {
-	conditions := []Condition{
-		{
-			ConditionName: "keyword",
-			Value:         "",
-		},
-		{
-			ConditionName: "user_id",
-			Value:         "",
-		},
-	}
-
-	result, err := getJoinedConditions(conditions)
-
-	assert.Equal(t, "", result)
-	assert.Equal(t, "could not join conditions", err.Error())
-}
